+++ src/main/java/roomescape/auth/resolver/LoginUserArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.global;
+package roomescape.auth.resolver;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
