+++ src/main/java/roomescape/global/config/WebMvcConfiguration.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.global.LoginUserArgumentResolver;
+import roomescape.auth.resolver.LoginUserArgumentResolver;
 
 @Configuration
 public class WebMvcConfiguration implements WebMvcConfigurer {
