+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -45,7 +45,8 @@ public class ReservationController {
     }
 
     @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> delete(@LoginUser Member member, @PathVariable("id") @Min(1) long reservationMemberId) {
+    public ResponseEntity<Void> delete(@LoginUser Member member,
+                                       @PathVariable("id") @Min(1) long reservationMemberId) {
         reservationService.deleteMemberReservation(member, reservationMemberId);
         return ResponseEntity.noContent().build();
     }
